engine: Move get_fastest_supported_compression_type into engine_rocks
