+++ components/engine/src/rocks/util/mod.rs
@@ -6,29 +6,13 @@ use std::sync::Arc;
 
 use crate::{Error, Result};
 use rocksdb::load_latest_options;
-use rocksdb::rocksdb::supported_compression;
 use rocksdb::{
-    CColumnFamilyDescriptor, ColumnFamilyOptions, DBCompressionType, DBOptions, Env,
+    CColumnFamilyDescriptor, ColumnFamilyOptions, DBOptions, Env,
     SliceTransform, DB,
 };
 
 pub use crate::rocks::CFHandle;
 use engine_traits::CF_DEFAULT;
-
-// Zlib and bzip2 are too slow.
-const COMPRESSION_PRIORITY: [DBCompressionType; 3] = [
-    DBCompressionType::Lz4,
-    DBCompressionType::Snappy,
-    DBCompressionType::Zstd,
-];
-
-pub fn get_fastest_supported_compression_type() -> DBCompressionType {
-    let all_supported_compression = supported_compression();
-    *COMPRESSION_PRIORITY
-        .iter()
-        .find(|c| all_supported_compression.contains(c))
-        .unwrap_or(&DBCompressionType::No)
-}
 
 pub fn get_cf_handle<'a>(db: &'a DB, cf: &str) -> Result<&'a CFHandle> {
     let handle = db
