+++ components/engine_rocks/src/sst.rs
@@ -14,12 +14,12 @@ use rocksdb::ExternalSstFileInfo as RawExternalSstFileInfo;
 use rocksdb::DB;
 use rocksdb::{ColumnFamilyOptions, SstFileReader};
 use rocksdb::{Env, EnvOptions, SequentialFile, SstFileWriter};
+use rocksdb::rocksdb::supported_compression;
 use std::rc::Rc;
 use std::sync::Arc;
 // FIXME: Move RocksSeekKey into a common module since
 // it's shared between multiple iterators
 use crate::engine_iterator::RocksSeekKey;
-use engine::rocks::util::get_fastest_supported_compression_type;
 use std::path::PathBuf;
 
 impl SstExt for RocksEngine {
@@ -246,6 +246,21 @@ impl ExternalSstFileInfo for RocksExternalSstFileInfo {
     fn num_entries(&self) -> u64 {
         self.0.num_entries()
     }
+}
+
+// Zlib and bzip2 are too slow.
+const COMPRESSION_PRIORITY: [DBCompressionType; 3] = [
+    DBCompressionType::Lz4,
+    DBCompressionType::Snappy,
+    DBCompressionType::Zstd,
+];
+
+fn get_fastest_supported_compression_type() -> DBCompressionType {
+    let all_supported_compression = supported_compression();
+    *COMPRESSION_PRIORITY
+        .iter()
+        .find(|c| all_supported_compression.contains(c))
+        .unwrap_or(&DBCompressionType::No)
 }
 
 #[cfg(test)]
